requirements.txt
discord.py
python-dotenv
supabase
pytz
requests
openai
coverage